# code/backend/metiche/rest/server/list_agent.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"fmt"
	"log"
	"net/http"

	"go.einride.tech/aip/filtering"
	"go.einride.tech/aip/ordering"

	agentmodule "github.com/mklfarha/metiche/backend/core/module/agent"
	"github.com/mklfarha/metiche/backend/core/module/agent/types"
	main_entity "github.com/mklfarha/metiche/backend/entity/agent"
)

func (h *handlers) ListAgent(w http.ResponseWriter, r *http.Request) {
	defaultPageSize := int32(10)
	maxPageSize := int32(100)

	lp, err := parseListParams(r, defaultPageSize, maxPageSize)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	offset, err := parsePageTokenOffset(lp.pageToken)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid page token: %w", err))
		return
	}

	declarations := agentDeclarations()
	filter, err := filtering.ParseFilter(lp, declarations)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid filter expression: %w", err))
		return
	}

	orderBy, err := ordering.ParseOrderBy(lp)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid order_by expression: %w", err))
		return
	}

	req := types.ListRequest{
		Offset:                offset,
		PageSize:              lp.GetPageSize(),
		Filter:                filter,
		FilteringDeclarations: declarations,
		EnumValues:            agentEnumValues(),
		OrderBy:               orderBy,
		IncludeFields:         lp.includeFields,
		ExcludeFields:         lp.excludeFields,
	}

	var result types.ListResponse
	if lp.skipCache {
		result, err = h.core.Agent().List(r.Context(), req, agentmodule.WithSkipCache())
	} else {
		result, err = h.core.Agent().List(r.Context(), req)
	}
	if err != nil {
		// A filter or order_by the clause builder cannot express is the caller's
		// error, not the server's: statusForError turns it into a 400.
		writeProblem(w, statusForError(err), err)
		return
	}

	resp := ListEnvelope[main_entity.Agent]{
		Items: result.Agent,
	}
	if result.HasNextPage {
		resp.NextPageToken = encodePageToken(offset + int64(lp.GetPageSize()))
	}

	writeJSON(w, http.StatusOK, resp)
}

func agentDeclarations() *filtering.Declarations {
	declarations, err := filtering.NewDeclarations(
		filtering.DeclareStandardFunctions(),
		filtering.DeclareIdent("true", filtering.TypeBool),
		filtering.DeclareIdent("false", filtering.TypeBool),

		// agent

		filtering.DeclareIdent("id", filtering.TypeString),

		filtering.DeclareIdent("key", filtering.TypeString),

		filtering.DeclareIdent("label", filtering.TypeString),

		filtering.DeclareIdent("client_kind", filtering.TypeString),

		filtering.DeclareIdent("client_key", filtering.TypeString),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in agentEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("status", filtering.TypeString),
		filtering.DeclareIdent("AGENT_STATUS_INVALID", filtering.TypeString),
		filtering.DeclareIdent("AGENT_STATUS_ACTIVE", filtering.TypeString),
		filtering.DeclareIdent("AGENT_STATUS_RETIRED", filtering.TypeString),

		filtering.DeclareIdent("last_seen_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("created_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("updated_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("account_uuid", filtering.TypeString),
	)
	if err != nil {
		log.Printf("error creating declarations for agent: %v", err)
	}
	return declarations
}

// agentEnumValues maps each enum filter field to every accepted
// spelling of its values and the number the column stores. REST generates no
// protobuf enum types (an app built with --api rest has no proto package at
// all), so this table — not the filter declarations — is what makes an enum
// filter resolvable here.
func agentEnumValues() map[string]map[string]int64 {
	return map[string]map[string]int64{

		"status": {
			"invalid":              0,
			"AGENT_STATUS_INVALID": 0,
			"active":               1,
			"AGENT_STATUS_ACTIVE":  1,
			"retired":              2,
			"AGENT_STATUS_RETIRED": 2,
		},
	}
}