# code/backend/metiche/rest/server/openapi.yaml
...
openapi: 3.1.0
info:
    title: Metiche
    version: 1.0.0
servers:
    - url: http://localhost:8080/v1
paths:
...
    /agents:
        get:
            tags:
                - Agent
            summary: List agents
            operationId: listAgent
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of agents
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - Agent
            summary: Create Agent
            operationId: createAgent
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Agent'
                required: true
            responses:
                "201":
                    description: Created Agent
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Agent'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /plans/{id}:
        get:
            tags:
                - Plan
            summary: Get Plan by ID
            operationId: getPlan
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Plan'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        patch:
            tags:
                - Plan
            summary: Partially update Plan
            operationId: patchPlan
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Plan'
                required: true
            responses:
                "200":
                    description: Successful update
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Plan'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Version conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        delete:
            tags:
                - Plan
            summary: Delete Plan by ID
            operationId: deletePlan
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
components:
    schemas:
...
        Agent:
            properties:
                account_uuid:
                    format: uuid
                    type: string
                client_key:
                    type: string
                client_kind:
                    nullable: true
                    type: string
                created_at:
                    format: date-time
                    type: string
                id:
                    format: uuid
                    type: string
                key:
                    type: string
                label:
                    type: string
                last_seen_at:
                    format: date-time
                    nullable: true
                    type: string
                status:
                    $ref: '#/components/schemas/AgentStatus'
                updated_at:
                    format: date-time
                    type: string
            required:
                - id
                - key
                - label
                - client_key
                - status
                - created_at
                - updated_at
                - account_uuid
            type: object
        AgentList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/Agent'
                    type: array
                next_page_token:
                    type: string
            type: object
        AgentStatus:
            enum:
                - invalid
                - active
                - retired
            type: string
...
        Plan:
            properties:
                created_at:
                    format: date-time
                    type: string
                description:
                    nullable: true
                    type: string
                id:
                    format: uuid
                    type: string
                is_instance_default:
                    type: boolean
                key:
                    type: string
                max_concurrent_agents:
                    format: int64
                    nullable: true
                    type: integer
                max_members:
                    format: int64
                    nullable: true
                    type: integer
                max_projects:
                    format: int64
                    nullable: true
                    type: integer
                name:
                    type: string
                retention_days:
                    format: int64
                    nullable: true
                    type: integer
                sort_order:
                    format: int64
                    type: integer
                status:
                    $ref: '#/components/schemas/RecordStatus'
                updated_at:
                    format: date-time
                    type: string
            required:
                - id
                - key
                - name
                - is_instance_default
                - sort_order
                - status
                - created_at
                - updated_at
            type: object
...
        Problem:
            properties:
                detail:
                    type: string
                instance:
                    type: string
                status:
                    type: integer
                title:
                    type: string
                type:
                    type: string
            type: object
...
        RecordStatus:
            enum:
                - invalid
                - active
                - inactive
            type: string
...